requirements.txt
opencv-python
pillow
easydict
numpy
matplotlib